file struct  modified and add event ui started
