+++ frontend/package.json
@@ -7,8 +7,12 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
     "@headlessui/react": "^1.6.1",
     "@heroicons/react": "^1.0.6",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.9.1",
     "axios": "^0.27.2",
     "date-fns": "^2.28.0",
     "next": "latest",
